Updated the codemod to simply append after the last import
