+++ scripts/component-generator/__test__/run-codemods.spec.js
@@ -2,4 +2,9 @@ jest.autoMockOff();
 const defineTest = require('jscodeshift/dist/testUtils').defineTest;
 
 // This will run the transform on the `../__testfixtures__/stories.input.js` file
-defineTest(__dirname, 'src/transform/stories-file', { ComponentName: 'MyNewComponent' }, 'stories');
+defineTest(
+  __dirname,
+  'src/transforms/stories-file',
+  { ComponentName: 'MyNewComponent' },
+  'stories',
+);
