+++ scripts/component-generator/__testfixtures__/stories.input.js
@@ -13,9 +13,5 @@ import '../stories/NewIcons'; // 1.4 Icons
 import '../stories/Collapse.story';
 import '../stories/Counter/index.story';
 
-// Generated components using the component generator
-import '../stories/SomePreviouslyGeneratedComponent/index.story';
-import '../stories/AnotherSomePreviouslyGeneratedComponent/index.story';
-
 // Styling
 import '../stories/Typography/TypographyClassesStory';
