+++ scripts/component-generator/src/run-codemods.js
@@ -3,27 +3,35 @@
 const path = require('path');
 const { exec } = require('child_process');
 const utils = require('./utils');
+const logger = require('./logger');
 
 const runTransform = (transformName, file, { ComponentName }) => {
   return new Promise((resolve, reject) => {
     const transformPath = path.join(__dirname, 'transforms', transformName);
-    const execProc = exec(
-      `jscodeshift -t ${transformPath} ${file} --ComponentName=${ComponentName}`,
+    const pathToExecutable = path.join(
+      __dirname,
+      '../../../node_modules/.bin/jscodeshift',
     );
 
-    execProc.stdout.on('data', data => console.log(data.toString()));
-    execProc.stderr.on('data', data => console.log(data.toString()));
-    execProc.on(
-      'exit',
-      code => console.log('Done with code:', code.toString()) || resolve(),
+    const execProc = exec(
+      `${pathToExecutable} -t ${transformPath} ${file} --ComponentName=${ComponentName}`,
     );
+
+    execProc.stderr.on('data', data => {
+      logger.error(
+        `Error while running codemod ${transformName}: ${data.toString}`,
+      );
+      reject(data.toString());
+    });
+
+    execProc.on('exit', resolve);
   });
 };
 
 module.exports = async ({ ComponentName }) => {
   // Run the stories-file transform
   await runTransform(
-    'stories-file',
+    'stories-file.js',
     utils.getDestinationPath('.storybook/stories.js'),
     { ComponentName },
   );
